add digits and some punctuation
The font includes glyphs only for English alphabet letters and the space.

Pages on the Animal Well Wiki [1] include other characters which aren't
included in the font:

  - digits, like in article "65th Egg"
    https://animalwell.wiki.gg/wiki/65th_Egg

  - colon, which are used to separate the namespace from the page name
    in a full page name, like "Talk:Main Page"
    https://animalwell.wiki.gg/wiki/Talk:Main_Page

  - parentheses, like in file page "File:2024-05-11 (2).png"
    https://animalwell.wiki.gg/index.php?title=File:2024-05-11_(2).png&redirect=no

  - forward slash, which is used in titles of subpages and some
    special pages:
    https://animalwell.wiki.gg/wiki/Template:Documentation/doc
    https://animalwell.wiki.gg/wiki/Special:PrefixIndex/Egg

Add the following characters to the font:

    0 1 2 3 4 5 6 7 8 9 . : , ; ( / \ ) - + [ ]

I used the font editor FontForge [2] to add these.  Although I am
familiar with some technical details about how fonts work, I am by no
means a graphical designer -- this was my first ever attempt at making a
font.  The majority of the glyphs were constructred simply by combining
elements of existing glyphs of the letters.

I don't like how the digits zero and five turned out.  For the digit
zero I made a so called "slashed zero" [3] based on the letter O, but
the "slash" is very thick, which makes the middle of the zero too
"busy"/"dark".  For the digit five, I used the letter "S" as the base,
which doesn't look very well on the left side on one of the bends:

             ****
             *
    here --> ***
                *
                *
             ***

Nevertheless, I think this should be good enough for now to make the
page names readable on the wiki.

[1] https://animalwell.wiki.gg/
[2] https://fontforge.org
[3] https://en.wikipedia.org/wiki/Slashed_zero